Move classes around and rename module
