+++ readthedocs/notifications/__init__.py
@@ -1,4 +1,4 @@
-from .library import Notification
+from .notification import Notification
 from .backends import send_notification
 
 default_app_config = 'readthedocs.notifications.apps.NotificationsAppConfig'
